requirements.txt
pyyaml
mysql-connector-python
coloredlogs
telethon
distutils2-py3
aiohttp
bs4
